requirements.txt
apt-clone==0.2.1
apt-xapian-index==0.47
apturl==0.5.2
blinker==1.3
Brlapi==0.6.4
chardet==2.3.0
command-not-found==0.3
cryptography==1.2.3
defer==1.0.6
dirspec==13.10
Django==1.11.6
django-cors-headers==2.1.0
djangorestframework==3.7.1
httplib2==0.9.1
idna==2.0
louis==2.6.4
lxml==3.5.0
Mako==1.0.3
MarkupSafe==0.23
oauthlib==1.0.3
onboard==1.2.0
oneconf==0.3.9
pexpect==4.0.1
Pillow==3.1.2
piston-mini-client==0.7.5
psutil==3.4.2
psycopg2==2.7.3.2
ptyprocess==0.5
pyasn1==0.1.9
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0
Pygments==2.2.0
pygobject==3.20.0
PyICU==1.9.2
PyJWT==1.3.0
python-apt==1.1.0b1
python-debian==0.1.27
python-xapp==1.0.0
pytz==2017.3
pyxdg==0.25
reportlab==3.3.0
requests==2.9.1
screen-resolution-extra==0.0.0
sessioninstaller==0.0.0
setproctitle==1.1.8
six==1.10.0
system-service==0.3
thin-client-config-agent==0.8
ubuntu-drivers-common==0.0.0
ufw==0.35
urllib3==1.13.1
virtkey==0.63.0
xkit==0.0.0
